Road to better maintainability
